split into 2 files, one of which abstracts away the permutation
